связь поста с пользователем при создании, обновлении и удалении

+++ src/post/post.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -13,13 +17,14 @@ export class PostService {
     private repository: Repository<PostEntity>,
   ) {}
 
-  create(dto: CreatePostDto) {
+  create(dto: CreatePostDto, userId: number) {
     const firstParagraph = dto.body.find((obj) => obj.type === 'paragraph')
       ?.data?.text;
     return this.repository.save({
       title: dto.title,
       body: dto.body,
       tags: dto.tags,
+      user: { id: userId },
       description: firstParagraph || '',
     });
   }
@@ -42,6 +47,7 @@ export class PostService {
 
   async search(dto: SearchPostDto) {
     const qb = this.repository.createQueryBuilder('p');
+    qb.leftJoinAndSelect('p.user', 'user');
     qb.limit(dto.limit || 0);
     qb.take(dto.limit || 10);
     if (dto.views) {
@@ -79,7 +85,7 @@ export class PostService {
     return this.repository.findOne({ where: { id } });
   }
 
-  async update(id: number, dto: UpdatePostDto) {
+  async update(id: number, dto: UpdatePostDto, userId: number) {
     const find = await this.repository.findOne({ where: { id } });
     if (!find) {
       throw new NotFoundException('Cтатья не найдена');
@@ -92,14 +98,18 @@ export class PostService {
       title: dto.title,
       body: dto.body,
       tags: dto.tags,
+      user: { id: userId },
       description: firstParagraph || '',
     });
   }
 
-  async remove(id: number) {
+  async remove(id: number, userId: number) {
     const find = await this.repository.findOne({ where: { id } });
     if (!find) {
       throw new NotFoundException('Cтатья не найдена');
+    }
+    if (find.user.id !== userId) {
+      throw new ForbiddenException('Нет доступа к статье');
     }
     return this.repository.delete(id);
   }
